Fix bug in register_map.sv
Don't set the ready bit when an instruction commits to the register file
on retire if the tag of the retiring instruction doesn't match the tag
of the instruction that will produce the newest value for that register.

This is to prevent newly dispatched instructions from reading incorrect (outdated)
register data.