Major overhaul: Code re-factoring & cleanup
- Added procyon_ff (flip-flop) and procyon_srff (SR flip-flop) modules to be used for any flops in the design
- Added procyon_onehot2binary and procyon_binary2onehot modules to convert between one-hot vectors and the bit index number
- Added procyon_priority_picker, a generic priority mux used for picking entries in buffers in the design
- re-factoring part 1: All buffers and queues now instantiate an "entry" module representing a single entry
of that buffer/queue. This helps separate out the per-entry logic for better readability and maintainability
- re-factoring part 2: Got rid of unnecessary signals in the design, cleaned up spaces in the code
- re-factoring part 3: Used the new procyon_ff/procyon_srff/procyon_onehot2binary/procyon_binary2onehot/procyon_priority_picker
everywhere in the design that they are needed. Made the FIFOs and ROB use a counter and head/tail logic wrap around when it reaches
it's end. This change now allows any buffers/queues in the design to be non-power-of-2 sized.
- re-factoring part 4: The dispatch pipeline was actually 3 cycles, this needed to be fixed to happen over 2 cycles. Fixed a bug
in the dispatch stage relating to the i_rs_stall and i_rob_stall signal being asserted on the same cycle causing an instruction
to not be enqueued in the reservation station.
- re-factoring part 5: Cleaned up the MHQ pipeline. It occurs over two cycles, "lookup" and "update"