procyon(core): Reserve entries in the ROB and RS in the same cycle in the dispatch stage
-- Doing this removes some inconsistencies when either the RS and ROB stalls during dispatch.
The core used to just reserve an entry in the ROB and fill the entry in the next cycle while the RS was allocated
in the next cycle which meant the output registers to the RS could not be modified while it was stalled whereas the ROB
already enqueued the instruction. This change makes it all or nothing, either we can reserve entries in both the ROB and
RS otherwise we stall.