procyon(core,lib): Add procyon_rr_picker and update procyon_rs_switch
This change makes procyon_rs_switch more resilient to changes to the number and layout of reservation stations.
Instead of hardcoding the connections, each RS will send a one-hot encoding of the class of ops they support and the
switch will pick 1 from a number of RS via a round robin arbitration scheme. This change adds a new module, procyon_rr_picker,
that implements the round robin arbitration.

In addition, the ROB is updated so that it no longer needs the i_rs_stall signal.