procyon(core): Fix a bug in the LSU when a retiring store follows a younger conflicting load in back-to-back cycles
Here is the sequence that resulted in this bug:

- A load is waiting in the reservation station. It is ready to be issued to the LSU and will do so in the next cycle.
At this point the load is not in the LQ and will be added in the DT/LQA cycle (i.e. 2 cycles later).
- In the same cycle the ROB signals the SQ to retire a store. The SQ will ack the ROB on the next cycle and mark it as retirable in the SQ.
- The load is now in the LSU_AD stage. The SQ has marked the store as retirable and is preparing to issue the store to the LSU in the next cycle.
- The load is now in the LSU_DT/LQA stage. The store has launched into the LSU_AD stage.
At this point the LQ is looked up for conflicting loads which are marked mis-speculated.

This last step is a problem because there is no conflict check on the newly allocating load when the store is retiring,
thus the load is allocated into the LQ and the mis-speculated bit is set to 0.